Map update, clean up

+++ Crossing Game/Car.cpp
@@ -31,7 +31,7 @@ Car::Car(int x, int y, int max_x, int max_y) {
 	length = 0;
 	for (wstring& s : sprite) {
 		if ((int)s.size() > length) {
-			length = s.size();
+			length = (int)s.size();
 		}
 	}
 	empty.append(speed, L' ');
